Migrate more dictionary-related code
